refactor(dataobj): move logs decoding to logs package
This change requires duplicating encoded metrics in the section
packages. An investigation on how to deduplicate these (such as moving
them to dataset) will be done following the completion of the overall
API refactor.
