+++ pkg/dataobj/internal/encoding/metrics.go
@@ -9,14 +9,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/grafana/loki/v3/pkg/dataobj/internal/metadata/datasetmd"
 	"github.com/grafana/loki/v3/pkg/dataobj/internal/metadata/filemd"
-	"github.com/grafana/loki/v3/pkg/dataobj/internal/metadata/logsmd"
-	"github.com/grafana/loki/v3/pkg/dataobj/internal/result"
-)
-
-var (
-	logsSectionLabels = prometheus.Labels{"section": SectionTypeLogs.String()}
 )
 
 // Metrics instruments encoded data objects.
@@ -24,23 +17,6 @@ type Metrics struct {
 	sectionsCount       prometheus.Histogram
 	fileMetadataSize    prometheus.Histogram
 	sectionMetadataSize *prometheus.HistogramVec
-
-	datasetColumnMetadataSize      prometheus.Histogram
-	datasetColumnMetadataTotalSize prometheus.Histogram
-
-	datasetColumnCount             prometheus.Histogram
-	datasetColumnCompressedBytes   *prometheus.HistogramVec
-	datasetColumnUncompressedBytes *prometheus.HistogramVec
-	datasetColumnCompressionRatio  *prometheus.HistogramVec
-	datasetColumnRows              *prometheus.HistogramVec
-	datasetColumnValues            *prometheus.HistogramVec
-
-	datasetPageCount             *prometheus.HistogramVec
-	datasetPageCompressedBytes   *prometheus.HistogramVec
-	datasetPageUncompressedBytes *prometheus.HistogramVec
-	datasetPageCompressionRatio  *prometheus.HistogramVec
-	datasetPageRows              *prometheus.HistogramVec
-	datasetPageValues            *prometheus.HistogramVec
 }
 
 // NewMetrics creates a new set of metrics for encoding.
@@ -70,132 +46,6 @@ func NewMetrics() *Metrics {
 			Name:      "section_metadata_size",
 			Help:      "Distribution of metadata size per encoded section.",
 		}, []string{"section"}),
-
-		datasetColumnMetadataSize: newNativeHistogram(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_metadata_size",
-			Help:      "Distribution of column metadata size per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}),
-
-		datasetColumnMetadataTotalSize: newNativeHistogram(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_metadata_total_size",
-			Help:      "Distribution of metadata size across all columns per encoded section.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}),
-
-		datasetColumnCount: newNativeHistogram(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_count",
-			Help:      "Distribution of column counts per encoded dataset section.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}),
-
-		datasetColumnCompressedBytes: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_compressed_bytes",
-			Help:      "Distribution of compressed bytes per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetColumnUncompressedBytes: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_uncompressed_bytes",
-			Help:      "Distribution of uncompressed bytes per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetColumnCompressionRatio: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_compression_ratio",
-			Help:      "Distribution of compression ratio per encoded dataset column. Not reported when compression is disabled.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type", "compression_type"}),
-
-		datasetColumnRows: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_rows",
-			Help:      "Distribution of row counts per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetColumnValues: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_column_values",
-			Help:      "Distribution of value counts per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetPageCount: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_count",
-			Help:      "Distribution of page count per encoded dataset column.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetPageCompressedBytes: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_compressed_bytes",
-			Help:      "Distribution of compressed bytes per encoded dataset page.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetPageUncompressedBytes: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_uncompressed_bytes",
-			Help:      "Distribution of uncompressed bytes per encoded dataset page.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetPageCompressionRatio: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_compression_ratio",
-			Help:      "Distribution of compression ratio per encoded dataset page. Not reported when compression is disabled.",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type", "compression_type"}),
-
-		datasetPageRows: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_rows",
-			Help:      "Distribution of row counts per encoded dataset page",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
-
-		datasetPageValues: newNativeHistogramVec(prometheus.HistogramOpts{
-			Namespace: "loki_dataobj",
-			Subsystem: "encoding",
-			Name:      "dataset_page_values",
-			Help:      "Distribution of value counts per encoded dataset page",
-
-			ConstLabels: logsSectionLabels, // Logs section is the only one that uses this metric.
-		}, []string{"column_type"}),
 	}
 }
 
@@ -221,20 +71,6 @@ func (m *Metrics) Register(reg prometheus.Registerer) error {
 	errs = append(errs, reg.Register(m.sectionsCount))
 	errs = append(errs, reg.Register(m.fileMetadataSize))
 	errs = append(errs, reg.Register(m.sectionMetadataSize))
-	errs = append(errs, reg.Register(m.datasetColumnMetadataSize))
-	errs = append(errs, reg.Register(m.datasetColumnMetadataTotalSize))
-	errs = append(errs, reg.Register(m.datasetColumnCount))
-	errs = append(errs, reg.Register(m.datasetColumnCompressedBytes))
-	errs = append(errs, reg.Register(m.datasetColumnUncompressedBytes))
-	errs = append(errs, reg.Register(m.datasetColumnCompressionRatio))
-	errs = append(errs, reg.Register(m.datasetColumnRows))
-	errs = append(errs, reg.Register(m.datasetColumnValues))
-	errs = append(errs, reg.Register(m.datasetPageCount))
-	errs = append(errs, reg.Register(m.datasetPageCompressedBytes))
-	errs = append(errs, reg.Register(m.datasetPageUncompressedBytes))
-	errs = append(errs, reg.Register(m.datasetPageCompressionRatio))
-	errs = append(errs, reg.Register(m.datasetPageRows))
-	errs = append(errs, reg.Register(m.datasetPageValues))
 	return errors.Join(errs...)
 }
 
@@ -243,20 +79,6 @@ func (m *Metrics) Unregister(reg prometheus.Registerer) {
 	reg.Unregister(m.sectionsCount)
 	reg.Unregister(m.fileMetadataSize)
 	reg.Unregister(m.sectionMetadataSize)
-	reg.Unregister(m.datasetColumnMetadataSize)
-	reg.Unregister(m.datasetColumnMetadataTotalSize)
-	reg.Unregister(m.datasetColumnCount)
-	reg.Unregister(m.datasetColumnCompressedBytes)
-	reg.Unregister(m.datasetColumnUncompressedBytes)
-	reg.Unregister(m.datasetColumnCompressionRatio)
-	reg.Unregister(m.datasetColumnRows)
-	reg.Unregister(m.datasetColumnValues)
-	reg.Unregister(m.datasetPageCount)
-	reg.Unregister(m.datasetPageCompressedBytes)
-	reg.Unregister(m.datasetPageUncompressedBytes)
-	reg.Unregister(m.datasetPageCompressionRatio)
-	reg.Unregister(m.datasetPageRows)
-	reg.Unregister(m.datasetPageValues)
 }
 
 // SectionObserver is invoked by [Metrics.Observe] for each section in the object.
@@ -286,13 +108,8 @@ func (m *Metrics) Observe(ctx context.Context, dec Decoder, onSection SectionObs
 
 		m.sectionMetadataSize.WithLabelValues(typ.String()).Observe(float64(calculateMetadataSize(section)))
 
-		switch typ {
-		case SectionTypeLogs:
-			errs = append(errs, m.observeLogsSection(ctx, dec.LogsDecoder(metadata, section)))
-		default:
-			if onSection != nil {
-				onSection(ctx, dec.SectionReader(metadata, section))
-			}
+		if onSection != nil {
+			onSection(ctx, dec.SectionReader(metadata, section))
 		}
 	}
 
@@ -310,58 +127,4 @@ func calculateMetadataSize(section *filemd.SectionInfo) uint64 {
 
 	// Fallback to the deprecated field.
 	return section.MetadataSize //nolint:staticcheck // MetadataSize is deprecated but still used as a fallback.
-}
-
-func (m *Metrics) observeLogsSection(ctx context.Context, dec LogsDecoder) error {
-	columns, err := dec.Columns(ctx)
-	if err != nil {
-		return err
-	}
-	m.datasetColumnCount.Observe(float64(len(columns)))
-
-	columnPages, err := result.Collect(dec.Pages(ctx, columns))
-	if err != nil {
-		return err
-	} else if len(columnPages) != len(columns) {
-		return fmt.Errorf("expected %d page lists, got %d", len(columns), len(columnPages))
-	}
-
-	// Count metadata sizes across columns.
-	{
-		var totalColumnMetadataSize int
-		for i := range columns {
-			columnMetadataSize := proto.Size(&logsmd.ColumnMetadata{Pages: columnPages[i]})
-			m.datasetColumnMetadataSize.Observe(float64(columnMetadataSize))
-			totalColumnMetadataSize += columnMetadataSize
-		}
-		m.datasetColumnMetadataTotalSize.Observe(float64(totalColumnMetadataSize))
-	}
-
-	for i, column := range columns {
-		columnType := column.Type.String()
-		pages := columnPages[i]
-		compression := column.Info.Compression
-
-		m.datasetColumnCompressedBytes.WithLabelValues(columnType).Observe(float64(column.Info.CompressedSize))
-		m.datasetColumnUncompressedBytes.WithLabelValues(columnType).Observe(float64(column.Info.UncompressedSize))
-		if compression != datasetmd.COMPRESSION_TYPE_NONE {
-			m.datasetColumnCompressionRatio.WithLabelValues(columnType, compression.String()).Observe(float64(column.Info.UncompressedSize) / float64(column.Info.CompressedSize))
-		}
-		m.datasetColumnRows.WithLabelValues(columnType).Observe(float64(column.Info.RowsCount))
-		m.datasetColumnValues.WithLabelValues(columnType).Observe(float64(column.Info.ValuesCount))
-
-		m.datasetPageCount.WithLabelValues(columnType).Observe(float64(len(pages)))
-
-		for _, page := range pages {
-			m.datasetPageCompressedBytes.WithLabelValues(columnType).Observe(float64(page.Info.CompressedSize))
-			m.datasetPageUncompressedBytes.WithLabelValues(columnType).Observe(float64(page.Info.UncompressedSize))
-			if compression != datasetmd.COMPRESSION_TYPE_NONE {
-				m.datasetPageCompressionRatio.WithLabelValues(columnType, compression.String()).Observe(float64(page.Info.UncompressedSize) / float64(page.Info.CompressedSize))
-			}
-			m.datasetPageRows.WithLabelValues(columnType).Observe(float64(page.Info.RowsCount))
-			m.datasetPageValues.WithLabelValues(columnType).Observe(float64(page.Info.ValuesCount))
-		}
-	}
-
-	return nil
 }
